messaging_app/requirement.txt
django
djangorestframework
django-cors-headers
celery
rabbitmq
drf-yasg
django-environ
mysqlclient
python-dotenv
drf-nested-routers
